Add Python 3.9.
This update also splits the installation of Python into two groups.
First the official Ubuntu packages are installed, and then the deadsnakes PPA is used.
This avoids conflicts when a given Python version is available from both sources, as is now the case with Python 3.8.

Additionally, the *-venv and *-distutils packages are now installed for all available versions.